requirements.txt
selenium==4.27.1
beautifulsoup4==4.12.3
aiohttp==3.11.10
requests==2.32.3
python-dotenv==1.0.1